Daily puzzles and larger word bank
Wordle uses 2 lists: the primary list that is the answer pool and a secondary list of allowed guesses that can never be answers.
The game will attempt to load the daily puzzle (calculated in the exact same way as the website) but if it's already been done today, loads a random word instead (there's a printed line if the daily puzzle is presented).