Histogram, version flag, cleanup, (secret emoji support 🤫)
The histogram tries to make the longest bar the same, max length every time and everything else to scale.

New -v version flag.

If the Win % rounds to a whole number, drop the decimal.

Quitting out reveals the word.

Secret emoji support! It's not really possible to programmatically determine if a terminal/font/config combination can display emoji. It'll default to false in the save file but a savvy/curious player can turn it on and see for themselves. I'm not responsible if your computer explodes.