Cleanup how stats are printed, better ctrl+c behavior around dailies
Printing stats was a mess. The code was duplicated between Hard and Normal mode, there were differences that shouldn't have been different, trimming the trailing decimal was converting "100.0" to "1". All streamlined and fixed.

If you started a daily puzzle and immediately quit, you shouldn't be punished by having the puzzle spoiled for you and being unable to play the puzzle on the next play. Quitting on the first guess won't reveal the word or reset your streak.

Cleaner labels and new lines for legibility.